requirements.txt
fastapi
uvicorn
httpx
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
